Add unit tests asserting form validation messages are displayed or hidden when expected.
Added unit tests to request-form.component.spec.ts to validate that validation error messages are not displayed before a user has edited a form field, are displayed for a field once a user has edited it to an invalid value (such as empty or invalid email format for the email field), and are not displayed when the user has entered a valid value in a field. Copied these and all other unit tests added to request-form.component.spec.ts to donation-form.component.spec.ts. All tests pass on both forms.