Fix error where form validation messages not appearing when expected.
Although form validation messages in the donation form and request form templates were set to appear when a form field had been edited and contained an invalid value (i.e. a user began to fill in and then cleared a field), these messages were not appearing when expected. Each form field previously contained a declaration of a reference variable with identical name, "#name." Now, each field of the template declares a reference variable unique to that field (e.g. "#phoneNumber"), and the div containing the error message uses the unique reference variable to check the state of the field. Error messages now appear as expected.